+++ christmas_app/accounts.py
@@ -7,7 +7,7 @@ from decouple import config as en_var # import the environment var
 usernames = {en_var('name_dad'):en_var('pass_dad'),en_var('name_mom'):en_var('pass_mom'),en_var('name_pa'):en_var('pass_pa'),en_var('name_mama'):en_var('pass_mama'),en_var('name_self'):en_var('pass_self') \
     ,"ADMIN":"admin"\
        ,en_var("name_papa"):en_var("pass_papa"), en_var("name_kwan"):en_var("pass_kwan"), en_var("name_wish"):en_var("pass_wish"), en_var("name_anna"):en_var('pass_anna'),en_var("name_ajT"):en_var("pass_ajT"),en_var('name_ajJ'):en_var("pass_ajJ") \
-        ,en_var("name_candice"):en_var("pass_candice"), en_var("name_addy"):en_var("pass_addy"), en_var("name_evander"):en_var("pass_evander"), en_var("name_noemi"):en_var("pass_noemi")}
+        ,en_var("name_candice"):en_var("pass_candice"), en_var("name_addy"):en_var("pass_addy"), en_var("name_evander"):en_var("pass_evander"), en_var("name_noemi"):en_var("pass_noemi"), en_var("name_grandma"):en_var("pass_grandma")}
 
 def create_dad():
     dad = None
@@ -54,7 +54,7 @@ def create_papa():
     try:
         papa = User(fname=en_var('name_papa'),password=generate_password_hash(usernames[en_var('name_papa')]).decode('utf-8'), alias='Papa')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database papa", category='danger')
     return papa
 
 def create_grandma():
@@ -62,7 +62,7 @@ def create_grandma():
     try:
         grandma = User(fname=en_var('name_grandma'),password=generate_password_hash(usernames[en_var('name_grandma')]).decode('utf-8'), alias='Grandma')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database grandma", category='danger')
     return grandma
 
 def create_kwan():
@@ -70,7 +70,7 @@ def create_kwan():
     try:
         kwan = User(fname=en_var('name_kwan'),password=generate_password_hash(usernames[en_var('name_kwan')]).decode('utf-8'), alias='Kwan')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database kwan", category='danger')
     return kwan
 
 def create_wish():
@@ -78,7 +78,7 @@ def create_wish():
     try:
         wish = User(fname=en_var('name_wish'),password=generate_password_hash(usernames[en_var('name_wish')]).decode('utf-8'), alias='Wish')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database wish", category='danger')
     return wish
 
 def create_anna():
@@ -86,7 +86,7 @@ def create_anna():
     try:
         anna = User(fname=en_var('name_anna'),password=generate_password_hash(usernames[en_var('name_anna')]).decode('utf-8'), alias='Anna')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database anna", category='danger')
     return anna
 
 def create_ajT():
@@ -94,7 +94,7 @@ def create_ajT():
     try:
         ajT = User(fname=en_var('name_ajT'),password=generate_password_hash(usernames[en_var('name_ajT')]).decode('utf-8'), alias='AjTrithep')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database ajt", category='danger')
     return ajT
 
 def create_ajJ():
@@ -102,7 +102,7 @@ def create_ajJ():
     try:
         ajJ = User(fname=en_var('name_ajJ'),password=generate_password_hash(usernames[en_var('name_ajJ')]).decode('utf-8'), alias='AjJade')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database ajj", category='danger')
     return ajJ
 
 def create_candice():
@@ -110,7 +110,7 @@ def create_candice():
     try:
         candice = User(fname=en_var('name_candice'),password=generate_password_hash(usernames[en_var('name_candice')]).decode('utf-8'), alias='Candice')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database candice", category='danger')
     return candice
 
 def create_addy():
@@ -118,7 +118,7 @@ def create_addy():
     try:
         addy = User(fname=en_var('name_addy'),password=generate_password_hash(usernames[en_var('name_addy')]).decode('utf-8'), alias='Addy')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database addy", category='danger')
     return addy
 
 def create_evander():
@@ -126,7 +126,7 @@ def create_evander():
     try:
         evander = User(fname=en_var('name_evander'),password=generate_password_hash(usernames[en_var('name_evander')]).decode('utf-8'), alias='Evander')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database evander", category='danger')
     return evander
 
 def create_noemi():
@@ -134,7 +134,7 @@ def create_noemi():
     try:
         noemi = User(fname=en_var('name_noemi'),password=generate_password_hash(usernames[en_var('name_noemi')]).decode('utf-8'), alias='Noemi')
     except:
-        flash("Couldn't add more row to database", category='danger')
+        flash("Couldn't add more row to database noemi", category='danger')
     return noemi
 
 def create_accounts():